Commit Texinfo output file and documentation index, as generated by:
  $ makeinfo histogram_ext.texi
  $ perl $MAXIMA_SRCDIR/doc/info/build_index.pl histogram_ext.info :utf8 > histogram_ext-index.lisp

where $MAXIMA_SRCDIR is the top directory of a Git repo or unpacked tarball.

Note that the ':utf8' argument is needed to enable build_index.pl to
handle UTF-8 characters in the .info file (originating from UTF-8 in
the .texi).

In a Maxima session, load("histogram_ext-index.lisp") loads the documentation index.
Then ? and ?? find the documentation items, e.g.: ?? bubble