Added a rtest file for histogram_ext and extened the comments in histogram_ext.mac
Copied rtest_histogram_previous.mac to rtest_histogram_ext.mac, as a comparision of features. Started to change the function names to histogram_ext and extended to function in case usefull options are missing. ... but not done yet, just started.
The goal is to mimic the interface closely, so none needs to relearn.

Extened the comments in histogram_ext.mac with most informations from .texi file.

Copied the function extract_options() from the descriptive.mac package.

Signed-off-by: Marcus Menzel <flareload@web.de>